refactor(documents): keep undo where it is used
The undo hook existed to be shared between the chat receipt and the
review toolbar. The receipt's undo button is gone, so it had one caller
and a comment describing a sharing that no longer happens. Fold it into
that caller, and inline the two-branch authorization helper alongside it.
